Massively simplified Core, added OccupancyMap and related functions. Added ConfigurationParser class to turn strings into occupancies. Added ThomasFermiDecorator, used with new function HartreeFocker::StartCore(). Added lattice resizing functionality to HFOperator (SpinorODE) so that potential size can be tested and adjusted using ExtendPotential(). Lots of the stuff downstream (MBPT, Basis, Configurations, ...) is now broken.

+++ Universal/Lattice.h
@@ -61,7 +61,7 @@ protected:
     const double* Calculate_Rpower(unsigned int k);
     
     /** Resizes the lattice such that NumPoints > min_size. */
-    virtual void ReSize(unsigned int min_size);
+    virtual void ReSize(double min_size);
 
 protected:
     unsigned int NumPoints;
